split animations manager
