+++ packages/react-native-reanimated/jest/resolver.js
@@ -8,6 +8,7 @@ const WEB_ONLY_IN_JEST = new Set([
   'UpdateLayoutAnimations',
   'useAnimatedRef',
   'useAnimatedStyle',
+  'animationsManager',
 ]);
 
 /** @type {import('jest-resolve').SyncResolver} */
